implement automatic database import/export to replace /var/lib/mysql volume
The application persisted its database using a Docker volume
mapped to a directory on the repository. This was space-inefficient
and caused problems with file permissions. Consequently, the
new volume contains a single .sql file which is a dump of the
entire database. It is imported automatically when the container
starts and updated automatically when the container stops.